various updates

+++ Assets/Scripts/ManaSlider.cs
@@ -7,13 +7,13 @@ public class ManaSlider : MonoBehaviour
 {
     public Slider _slider;
 
-    public void SetMaxMana(int mana)
+    public void SetMaxMana(float mana)
     {
         _slider.maxValue = mana;
         _slider.value = mana;
     }
 
-    public void SetMana(int mana)
+    public void SetMana(float mana)
     {
         _slider.value = mana;
     }
